+++ src/gamspy/_backend/backend.py
@@ -51,6 +51,8 @@ HEADER = [
 def backend_factory(
     container: Container,
     options: Options | None = None,
+    solver: str | None = None,
+    solver_options: dict | None = None,
     output: io.TextIOWrapper | None = None,
     backend: Literal["local", "engine", "neos"] = "local",
     client: EngineClient | NeosClient | None = None,
@@ -63,6 +65,8 @@ def backend_factory(
         return NEOSServer(
             container,
             options,  # type: ignore
+            solver,
+            solver_options,
             client,  # type: ignore
             output,
             model,
@@ -75,6 +79,8 @@ def backend_factory(
             container,
             client,  # type: ignore
             options,  # type: ignore
+            solver,
+            solver_options,
             output,
             model,
             load_symbols,
@@ -82,7 +88,15 @@ def backend_factory(
     elif backend == "local":
         from gamspy._backend.local import Local
 
-        return Local(container, options, output, model, load_symbols)
+        return Local(
+            container,
+            options,
+            solver,
+            solver_options,
+            output,
+            model,
+            load_symbols,
+        )
 
     raise ValidationError(
         f"`{backend}` is not a valid backend. Possible backends:"
@@ -96,12 +110,16 @@ class Backend(ABC):
         container: Container,
         model: Model,
         options: Options,
+        solver: str | None,
+        solver_options: dict | None,
         output: io.TextIOWrapper | None,
         load_symbols: list[Symbol] | None,
     ):
         self.container = container
         self.model = model
         self.options = options
+        self.solver = solver
+        self.solver_options = solver_options
         self.output = output
         self.load_symbols = load_symbols
         if load_symbols is not None: