+++ tests/integration/test_neos.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
 import os
+import sys
 import unittest
 
 import pandas as pd
@@ -137,7 +138,7 @@ class NeosSuite(unittest.TestCase):
         client = NeosClient(
             email=os.environ["NEOS_EMAIL"],
         )
-        transport.solve(backend="neos", client=client, solver="CONOPT")
+        transport.solve(backend="neos", client=client, solver="cplex")
 
         import math
 
@@ -191,6 +192,99 @@ class NeosSuite(unittest.TestCase):
         self.assertTrue("x" in container.data)
         x.setRecords(container["x"].records)
         self.assertTrue(x.records.equals(container["x"].records))
+
+    def test_solver_options(self):
+        # Set
+        i = Set(
+            self.m,
+            name="i",
+            records=["seattle", "san-diego"],
+            description="canning plants",
+        )
+        j = Set(
+            self.m,
+            name="j",
+            records=["new-york", "chicago", "topeka"],
+            description="markets",
+        )
+
+        # Data
+        a = Parameter(
+            self.m,
+            name="a",
+            domain=i,
+            records=self.capacities,
+            description="capacity of plant i in cases",
+        )
+        b = Parameter(
+            self.m,
+            name="b",
+            domain=j,
+            records=self.demands,
+            description="demand at market j in cases",
+        )
+        d = Parameter(
+            self.m,
+            name="d",
+            domain=[i, j],
+            records=self.distances,
+            description="distance in thousands of miles",
+        )
+        c = Parameter(
+            self.m,
+            name="c",
+            domain=[i, j],
+            description="transport cost in thousands of dollars per case",
+        )
+        c[i, j] = 90 * d[i, j] / 1000
+
+        # Variable
+        x = Variable(
+            self.m,
+            name="x",
+            domain=[i, j],
+            type="Positive",
+            description="shipment quantities in cases",
+        )
+
+        # Equation
+        supply = Equation(
+            self.m,
+            name="supply",
+            domain=i,
+            description="observe supply limit at plant i",
+        )
+        demand = Equation(
+            self.m,
+            name="demand",
+            domain=j,
+            description="satisfy demand at market j",
+        )
+
+        supply[i] = Sum(j, x[i, j]) <= a[i]
+        demand[j] = Sum(i, x[i, j]) >= b[j]
+
+        transport = Model(
+            self.m,
+            name="transport",
+            equations=self.m.getEquations(),
+            problem="LP",
+            sense=Sense.MIN,
+            objective=Sum((i, j), c[i, j] * x[i, j]),
+        )
+        client = NeosClient(
+            email=os.environ["NEOS_EMAIL"],
+        )
+        transport.solve(
+            output=sys.stdout,
+            solver="cplex",
+            solver_options={"aggfill": "11"},
+            backend="neos",
+            client=client,
+        )
+
+        with open(os.path.join(self.m.working_directory, "solve.log")) as file:
+            self.assertTrue(">>  aggfill 11" in file.read())
 
 
 def neos_suite():